[Decomp][ObjectMgr] Cohesion split: extract game teleport loaders to ObjectMgrGameTele.cpp
